requirements.txt
# Python requirements for this playbook

ansible == 2.9.5
boto == 2.49.0
boto3 == 1.11.11
netaddr == 0.7.19
openstacksdk == 0.41.0
pyOpenSSL == 19.0.0
